Fixed classifier for images

+++ app/src/main/java/com/example/emotionrecognition/Classifier.kt
@@ -15,7 +15,8 @@ class LinearSVC(context: Context) {
     }
 
     private val clf: Classifier
-    fun predict(features: DoubleArray): Int {
+
+    fun predict(features: FloatArray): Int {
         var classIdx = 0
         var classVal = Double.NEGATIVE_INFINITY
         var i = 0
